+++ examples/aws-lambda/lambda/graphql.ts
@@ -1,8 +1,10 @@
-import type { Handler } from '@aws-cdk/aws-lambda'
 import { createYoga, createSchema } from 'graphql-yoga'
-import { configure } from '@vendia/serverless-express'
+import { APIGatewayEvent, APIGatewayProxyResult, Context } from 'aws-lambda'
 
-const app = createYoga({
+const yoga = createYoga<{
+  event: APIGatewayEvent
+  lambdaContext: Context
+}>({
   graphqlEndpoint: '/graphql',
   landingPage: false,
   schema: createSchema({
@@ -20,7 +22,45 @@ const app = createYoga({
   }),
 })
 
-export const handler: Handler = configure({
-  app,
-  log: app.logger,
-})
+export async function handler(
+  event: APIGatewayEvent,
+  lambdaContext: Context,
+): Promise<APIGatewayProxyResult> {
+  const url = new URL(event.path, 'http://localhost')
+  if (event.queryStringParameters != null) {
+    for (const name in event.queryStringParameters) {
+      const value = event.queryStringParameters[name]
+      if (value != null) {
+        url.searchParams.set(name, value)
+      }
+    }
+  }
+
+  const response = await yoga.fetch(
+    url,
+    {
+      method: event.httpMethod,
+      headers: event.headers as HeadersInit,
+      body: event.body
+        ? Buffer.from(event.body, event.isBase64Encoded ? 'base64' : 'utf8')
+        : undefined,
+    },
+    {
+      event,
+      lambdaContext,
+    },
+  )
+
+  const responseHeaders: Record<string, string> = {}
+
+  response.headers.forEach((value, name) => {
+    responseHeaders[name] = value
+  })
+
+  return {
+    statusCode: response.status,
+    headers: responseHeaders,
+    body: await response.text(),
+    isBase64Encoded: false,
+  }
+}