+++ examples/error-handling/src/index.ts
@@ -3,4 +3,6 @@ import { yoga } from './yoga'
 
 // Start the server and explore http://localhost:4000/graphql
 const server = createServer(yoga)
-server.listen(4000)
+server.listen(4000, () => {
+  console.info('Server is running on http://localhost:4000/graphql')
+})