+++ examples/hackernews/src/main.ts
@@ -6,7 +6,9 @@ import { createContext } from './context'
 function main() {
   const yoga = createYoga({ schema, context: createContext })
   const server = createServer(yoga)
-  server.listen(4000)
+  server.listen(4000, () => {
+    console.info('Server is running on http://localhost:4000/graphql')
+  })
 }
 
 main()