+++ examples/hello-world/index.js
@@ -25,4 +25,6 @@ const yoga = createYoga({
 })
 
 const server = createServer(yoga)
-server.listen(4000)
+server.listen(4000, () => {
+  console.log('Server listening on http://localhost:4000/graphql')
+})