+++ examples/koa/src/app.ts
@@ -1,5 +1,6 @@
 import { createYoga, createSchema } from 'graphql-yoga'
 import Koa from 'koa'
+import { Readable } from 'stream'
 
 export function buildApp() {
   const app = new Koa()