+++ examples/netlify-edge/src/index.ts
@@ -1,0 +1,19 @@
+import { createYoga, createSchema } from 'graphql-yoga'
+
+const yoga = createYoga({
+  schema: createSchema({
+    typeDefs: /* GraphQL */ `
+      type Query {
+        greetings: String!
+      }
+    `,
+    resolvers: {
+      Query: {
+        greetings: () =>
+          'This is the `greetings` field of the root `Query` type',
+      },
+    },
+  }),
+})
+
+export default yoga