+++ examples/node-ts/__integration-tests__/node-ts.spec.ts
@@ -2,17 +2,17 @@ import { yoga } from '../src/yoga'
 
 describe('node-ts example integration', () => {
   it('should execute query', async () => {
-    const response = await yoga.fetch('/graphql?query={hello}')
+    const response = await yoga.fetch('http://yoga/graphql?query={greetings}')
 
     expect(response.status).toBe(200)
     expect(await response.text()).toMatchInlineSnapshot(
-      `"{"data":{"hello":"world"}}"`,
+      `"{"data":{"greetings":"This is the \`greetings\` field of the root \`Query\` type"}}"`,
     )
   })
 
   it('should have subscriptions disabled', async () => {
     const response = await yoga.fetch(
-      '/graphql?query=subscription{greetings}',
+      'http://yoga/graphql?query=subscription{greetings}',
       {
         headers: {
           Accept: 'text/event-stream',