+++ packages/graphql-yoga/src/cli/index.ts
@@ -72,7 +72,9 @@ export function graphqlYoga() {
       })
       const server = createServer(yoga)
       console.log(`Starting GraphQL Server`)
-      server.listen(4000)
+      server.listen(4000, () => {
+        console.info('Server is running on http://localhost:4000/graphql')
+      })
 
       registerTerminateHandler(() => {
         server.close()