+++ packages/graphiql/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @graphql-yoga/graphiql
 
+## 3.0.0-next.1
+
+### Patch Changes
+
+- [`64e06d74`](https://github.com/dotansimha/graphql-yoga/commit/64e06d74132a118f30b42b51c0e71abced0506a4) Thanks [@ardatan](https://github.com/ardatan)! - Fix execute/stop button
+
 ## 3.0.0-next.0
 
 ### Major Changes